refactor: extract eva-orchestrator helpers to separate module (#SD-MAN-INFRA-CORRECTIVE-ARCHITECTURE-GAP-020)
Extract 6 internal helper functions (buildResult, mergeArtifactOutputs,
loadStageTemplate, persistArtifacts, checkIdempotency, loadUpstreamArtifacts)
and constants (STATUS, FILTER_ACTION, ANALYSIS_DEPTH, selectAnalysisDepth)
from eva-orchestrator.js into eva-orchestrator-helpers.js.

Reduces main module from 966 LOC to 766 LOC while maintaining full
backward compatibility via re-exported _internal object.
